refactor(engine): extract daemon domain types to src/daemon/types.ts
Domain types (WorkflowTrigger, WorkflowRunResult variants, SessionSource,
AllocatedSession, ChildWorkflowRunResult, OrphanedSession, ReadFileState)
moved from workflow-runner.ts to a zero-dep types.ts. Every consumer that
previously imported WorkflowTrigger was forced to load AnthropicBedrock and
node:fs as side effects; now they load nothing but the type itself.

Re-exports in workflow-runner.ts maintain backward compat for the transition.
context-loader.ts and session-scope.ts now import from types.ts (correct
direction); trigger/ and v2/usecases/ likewise. The circular import note in
session-scope.ts is removed -- the circular dependency no longer exists.

377/377 tests pass. Build clean.
